+++ pages/components/bottom-bar.tsx
