+++ pages/components/faq.tsx
